「初級理解度チェック　問題２」の修正をコミット
