+++ src/main/java/org/example/Student.java
@@ -1,0 +1,13 @@
+package org.example;
+
+public class Student {
+
+  String name;
+  int score;
+
+  Student(String name, int score) {
+    this.name = name;
+    this.score = score;
+  }
+
+}
