+++ src/main/java/org/example/StudentList.java
@@ -1,0 +1,14 @@
+package org.example;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class StudentList {
+
+  public List<Student> students;
+
+  public StudentList() {
+    this.students = new ArrayList<>();
+  }
+
+}
